Generate Test Data for Sparse Matrices

This function creates a list of sparse matrices (K27, CUTAC, K4) with randomly generated binary values (0 or 1). The matrices simulate different datasets with overlapping and unique column names.

getTestData()

Value

A list of sparse matrices:

K27

A 10-row matrix with 5 columns labeled as Cell1 to Cell5.

CUTAC

A 10-row matrix with 5 columns labeled as Cell2, Cell3, Cell6, Cell7, Cell8.

K4

A 10-row matrix with 5 columns labeled as Cell3, Cell4, Cell7, Cell9, Cell10.

Details

Each matrix has 10 rows, with row names prefixed by "K27_", "CUTAC_", or "K4_", respectively.

Column names simulate cell identifiers, demonstrating a scenario where datasets share some cells but not others.

The function sets a random seed (set.seed(123)) to ensure reproducibility.

Examples

testData <- getTestData()
str(testData)  # Check the structure of the output
#> List of 3
#>  $ K27  :Formal class 'dgCMatrix' [package "Matrix"] with 6 slots
#>   .. ..@ i       : int [1:20] 3 5 6 7 0 1 2 4 6 1 ...
#>   .. ..@ p       : int [1:6] 0 4 9 13 17 20
#>   .. ..@ Dim     : int [1:2] 10 5
#>   .. ..@ Dimnames:List of 2
#>   .. .. ..$ : chr [1:10] "K27_1" "K27_2" "K27_3" "K27_4" ...
#>   .. .. ..$ : chr [1:5] "Cell1" "Cell2" "Cell3" "Cell4" ...
#>   .. ..@ x       : num [1:20] 1 1 1 1 1 1 1 1 1 1 ...
#>   .. ..@ factors : list()
#>  $ CUTAC:Formal class 'dgCMatrix' [package "Matrix"] with 6 slots
#>   .. ..@ i       : int [1:23] 0 5 6 8 1 2 5 0 5 6 ...
#>   .. ..@ p       : int [1:6] 0 4 7 12 18 23
#>   .. ..@ Dim     : int [1:2] 10 5
#>   .. ..@ Dimnames:List of 2
#>   .. .. ..$ : chr [1:10] "CUTAC_1" "CUTAC_2" "CUTAC_3" "CUTAC_4" ...
#>   .. .. ..$ : chr [1:5] "Cell2" "Cell3" "Cell6" "Cell7" ...
#>   .. ..@ x       : num [1:23] 1 1 1 1 1 1 1 1 1 1 ...
#>   .. ..@ factors : list()
#>  $ K4   :Formal class 'dgCMatrix' [package "Matrix"] with 6 slots
#>   .. ..@ i       : int [1:31] 1 2 4 5 9 0 1 2 5 6 ...
#>   .. ..@ p       : int [1:6] 0 5 12 15 23 31
#>   .. ..@ Dim     : int [1:2] 10 5
#>   .. ..@ Dimnames:List of 2
#>   .. .. ..$ : chr [1:10] "K4_1" "K4_2" "K4_3" "K4_4" ...
#>   .. .. ..$ : chr [1:5] "Cell3" "Cell4" "Cell7" "Cell9" ...
#>   .. ..@ x       : num [1:31] 1 1 1 1 1 1 1 1 1 1 ...
#>   .. ..@ factors : list()

# Accessing individual matrices
testData[[1]]  # K27 matrix
#> 10 x 5 sparse Matrix of class "dgCMatrix"
#>        Cell1 Cell2 Cell3 Cell4 Cell5
#> K27_1      .     1     .     .     .
#> K27_2      .     1     1     1     1
#> K27_3      .     1     .     .     1
#> K27_4      1     .     .     1     .
#> K27_5      .     1     .     1     .
#> K27_6      1     .     .     .     .
#> K27_7      1     1     1     .     .
#> K27_8      1     .     1     .     1
#> K27_9      .     .     .     .     .
#> K27_10     .     .     1     1     .
testData[[2]]  # CUTAC matrix
#> 10 x 5 sparse Matrix of class "dgCMatrix"
#>          Cell2 Cell3 Cell6 Cell7 Cell8
#> CUTAC_1      1     .     1     1     1
#> CUTAC_2      .     1     .     1     1
#> CUTAC_3      .     1     .     .     .
#> CUTAC_4      .     .     .     1     1
#> CUTAC_5      .     .     .     1     1
#> CUTAC_6      1     1     1     1     .
#> CUTAC_7      1     .     1     .     .
#> CUTAC_8      .     .     .     .     1
#> CUTAC_9      1     .     1     1     .
#> CUTAC_10     .     .     1     .     .
testData[[3]]  # K4 matrix
#> 10 x 5 sparse Matrix of class "dgCMatrix"
#>       Cell3 Cell4 Cell7 Cell9 Cell10
#> K4_1      .     1     .     1      1
#> K4_2      1     1     1     .      1
#> K4_3      1     1     .     1      1
#> K4_4      .     .     .     1      .
#> K4_5      1     .     .     1      1
#> K4_6      1     1     1     1      1
#> K4_7      .     1     1     1      1
#> K4_8      .     1     .     1      .
#> K4_9      .     .     .     1      1
#> K4_10     1     1     .     .      1